First version of the algorithm
Given a json file with the nodes of a dungeon, the algorithm will find the sequence of events on how to "beat" this dungeon.
A node can have requirements to be visited, once visited , if the node has any unlocks, you can unlock things in there, when a room is unlocked you can now visit it.